* gui_system : support des touches Home (curseur au début), End (curseur à la fin), Suppr (supprimer devant le curseur) dans les textbox * suppression de warnings (WinInventory, EntityFactory, Item, BitmapString) * nouvelles tiles, nouvelles tiles animées (chutes d'eau)

+++ src/misc/BitmapString.cpp
@@ -74,7 +74,7 @@ void BitmapString::RemoveChar(int position)
 	}
 
 	position = GetRealPosition(position);
-	if (position > (chars_.size() - 1))
+	if (position > (int) chars_.size() - 1)
 	{
 		position = chars_.size() - 1;
 	}
